requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
torch==2.4.1
torchvision==0.19.1
scikit-image==0.24.0
matplotlib==3.9.2
tqdm==4.66.5
pandas==2.2.3
PyYAML==6.0.2
streamlit==1.39.0
